validacion de empleados completado, falta validacion de employeeUpdate y assets

+++ controllers/employee_controller.js
@@ -1,6 +1,9 @@
 const model = require("../models/employee");
-const validarSolicitud = require("../middleware/validarEmployee")
+const {validationResult} = require("express-validator")
+
+// const {validarSolicitud} = require("../middleware/validarEmployee")
 const NotFoundError = require("../errors/NotFoundError")
+
 
 //--FILTRAO GENERAL--
 const getAllEmployees = async (req, res, next) => {
@@ -14,18 +17,22 @@ const getEmployeesById = async (req, res, next) => {
   try {
     const userId = req.params.id;
     const user = await model.getEmployeeByIdModel(userId);
-    console.log(user)
     if (!user) throw new NotFoundError('El empleado no existe') ; //validacion
     res.json({ menssage: user });
   } catch (error) {
-    console.log(error)
     next(error)
   }
 };
 
 //--CREAR EMPLEADO--
 const createEmployee = async (req, res) => {
-  validarSolicitud(req)
+
+  //Validacion de resultados de express validator: Me surgian muchos errores mandando a un archivo aparte
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const values = {...req.body}; //--Traemos valores del cuerpo
     const result = await model.createEmployeeModel(values);
@@ -37,13 +44,18 @@ const createEmployee = async (req, res) => {
 
 //--ACTUALIZAR EMPLEADO--
 const updateEmployee = async (req, res, next) => {
-  // validarSolicitud(req)
+
+  //Validacion de resultados de express validator: Me surgian muchos errores mandando a un archivo aparte
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
-    //validar si existe user id
     const userId = req.params.id;
-    const user = await model.getEmployeeByIdModel(userId);
-    if (!user) throw new NotFoundError('El empleado no existe') ; //validacion
-    const values = {...req.body};
+    const user = await model.getEmployeeByIdModel(userId); //valores viejos del usuario
+    if (!user) throw new NotFoundError('El empleado no existe'); //validacion
+    const values = {...req.body}; //valores nuevos del empleado
     const result = await model.updateEmployeeModel(user,values)
     res.json({result, messaje: `El usuario con el id ${userId} se actualizó exitosamente`})
   } catch (error) {
@@ -56,6 +68,7 @@ const updateEmployee = async (req, res, next) => {
 const deleteEmployee = async (req, res,next) => {
   try {
     const userId = req.params.eid;
+    if (!user) throw new NotFoundError('El empleado no existe');
     await model.deleteEmployeeModel(userId);
     res.json({ messaje: `Usuario con el Id: ${userId} eliminado` });
   } catch (error) {
